fix: switch back from esm to cjs
We want to use this package together w/ `tsc` and sometimes with `ts-node[-dev]` as well.

We've been debugging various issues when trying to implement this, and
the only thing that helped was using cjs instead of esm.

Currently we have a fork of this repo as a git submodule, with the
submitted PRs implemented in a separate branch, and use that directly
instead of an npm package - it works great, but this is far from ideal
and going back to cjs / using it in general means better compatibility
for _all_ of your consumers - both those using cjs and esm.
This does _not_ apply the other way around, and that's why it's an
issue for us that this package is with esm.

If this deems unacceptible, could you keep the things as-is,
but also create a separate file that uses cjs
which we could import from the npm package
something similar to this:

```js
import { getAttribute } from "xattr/cjs"
```

or similar?

---

see also:
https://twitter.com/Rich_Harris/status/1441068317930819589

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>